+++ src/flamenco-manager/flamenco/dashboard.go
@@ -108,8 +108,11 @@ func (rep *Reporter) sendStatusReport(w http.ResponseWriter, r *http.Request) {
 		}},
 		// 3: Project to just get what we need.
 		M{"$project": M{
-			"current_task_status":  "$_task.status",
-			"current_task_updated": "$_task.last_updated",
+			// To get the info from the task itself, swap out these two lines with the two lines below.
+			// "current_task_status":  "$_task.status",
+			// "current_task_updated": "$_task.last_updated",
+			"current_task_status":  1,
+			"current_task_updated": 1,
 			"address":              1,
 			"current_task":         1,
 			"last_activity":        1,